Fix 좋아요 10개 게시글 수정

+++ src/main/java/com/move/TripBalance/post/controller/response/TopFiveResponseDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TopFiveResponseDto {
-    private int heartNum;
+    private Long heartNum;
     private Long postId;
     private String title;
     private String img;
